requirements.txt
telethon==1.30.0
scikit-learn==1.5.1
joblib==1.3.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
fastapi
uvicorn[standard]
scikit-learn
pandas
requests
beautifulsoup4
lxml
python-multipart
tqdm
joblib
urllib3
